working back & front end
